+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
 - First version
+
+## 1.0.1
+
+- Supports `alignment` configuration in absolute position mode
+
+- Fix FFloatController status monitoring problem